% Generated by roxygen2 (4.0.1): do not edit by hand
\name{as.data.frame.OptPathDF}
\alias{as.data.frame.OptPathDF}
\title{Convert optimization path to data.frame.}
\usage{
\method{as.data.frame}{OptPathDF}(x, row.names = NULL, optional = FALSE,
  discretes.as.factor = FALSE, include.x = TRUE, include.y = TRUE,
  include.rest = TRUE, ...)
}
\arguments{
\item{x}{[\code{\link{OptPath}}]\cr
Optimization path.}

\item{row.names}{[\code{character}]\cr
Row names for result.
Default is none.}

\item{optional}{[any]\cr
Currently ignored.}

\item{discretes.as.factor}{[\code{logical(1)}]\cr
Represent discrete param columns as factors (or characters)?
Default is \code{FALSE}.}

\item{include.x}{[\code{logical(1)}]\cr
Include all input params?
Default is \code{TRUE}.}

\item{include.y}{[\code{logical(1)}]\cr
Include all y-columns?
Default is \code{TRUE}.}

\item{include.rest}{[\code{logical(1)}]\cr
Include all other columns?
Default is \code{TRUE}.}

\item{...}{[any] \cr
Currently ignored.}
}
\value{
[\code{data.frame}].
}
\description{
Convert optimization path to data.frame.
}

