% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkParamSet.R
\name{checkParamSet}
\alias{checkParamSet}
\title{Sanity check expressions of a parameter set.}
\usage{
checkParamSet(par.set, dict = NULL)
}
\arguments{
\item{par.set}{[\code{\link{ParamSet}}]\cr
Parameter set.}

\item{dict}{[\code{list} | \code{NULL}]\cr
If any of the parameters components is defined with an expression, this
list will be used to replace the values within the expressions.
The default is \code{NULL}.}
}
\value{
[\code{TRUE}].
}
\description{
Checks whether the default values of the numerical parameters are located
within the corresponding boundaries. In case of discrete parameters it
checks whether the values are a subset of the parameter's possible values.
}
\examples{
ps = makeParamSet(
  makeNumericParam("u", lower = expression(p)),
  makeIntegerParam("v", lower = 1, upper = expression(3 * p)),
  makeDiscreteParam("w", default = expression(z), values = c("a", "b")),
  makeDiscreteParam("x", default = "a", values = c("a", "b")),
  dictionary = c("p", "z")
)
checkParamSet(ps, dict = list(p = 3, z = "b"))
}

