% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkParVals.R
\name{checkParVals}
\alias{checkParVals}
\title{Check if a param value is technically valid.}
\usage{
checkParVals(par.set, x)
}
\arguments{
\item{par.set}{[\code{\link{ParamSet}}]\cr
Parameter set.}

\item{x}{[any] \cr
Single value to check.
For a parameter set this must be a list.}
}
\value{
Nothing. The function produces a meaningful exception
  if conditions are violated.
}
\description{
This is not a check for feasibility, e.g. checking of constraints, but whether parameters
have the correct names and data types, so that the parameter setting technically makes sense
for a \code{ParamSet}.
}

