% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dropParams.R
\name{dropParams}
\alias{dropParams}
\title{Drop Params from ParamSet by ids.}
\usage{
dropParams(par.set, drop)
}
\arguments{
\item{par.set}{\link{ParamSet}\cr
Parameter set.}

\item{drop}{(\code{character})\cr
\code{id}s of the \code{\link[=Param]{Param()}}s in the \code{\link[=ParamSet]{ParamSet()}} to drop from the ParamSet.}
}
\value{
[\code{\link[=ParamSet]{ParamSet()}}].
}
\description{
Drop Params from ParamSet by ids.
}
