% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluateParamSet.R
\name{evaluateParamSet}
\alias{evaluateParamSet}
\title{Evaluates expressions of a parameter set.}
\usage{
evaluateParamSet(par.set, envir = NULL)
}
\arguments{
\item{par.set}{[\code{\link{ParamSet}}]\cr
Parameter set.}

\item{envir}{[\code{list} | \code{NULL}]\cr
If any of the parameters components is defined with an expression, this
object will be used to replace the values within the expressions.
The default is \code{NULL}.}
}
\value{
[\code{TRUE}].
}
\description{
Evaluates the expressions of a parameter set for a given environment.
}
\examples{
ps = makeParamSet(
  makeNumericParam("u", lower=1),
  makeIntegerParam("v", lower=1, upper=2),
  makeDiscreteParam("w", values=1:2),
  makeLogicalParam("x"),
  makeDiscreteVectorParam("y", len=2, values=c("a", "b"))
)
}

