% Generated by roxygen2 (4.0.0): do not edit by hand
\name{extractParamSetFromSooFunction}
\alias{extractParamSetFromSooFunction}
\title{Extracts ParamSet from soobench function.}
\usage{
extractParamSetFromSooFunction(fn)
}
\arguments{
\item{fn}{[\code{\link[soobench]{soo_function}}]\cr
Source function.}
}
\value{
[\code{\link[ParamHelpers]{ParamSet}}]
  Parameter set of type \code{\link[ParamHelpers]{ParamSet}}.
}
\description{
Extracts ParamSet from soobench function.
}
\examples{
library(soobench)
fn = ackley_function(4)
par.set = extractParamSetFromSooFunction(fn)
print(par.set)
}

