% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateDesignOversampling.R
\name{generateDesignOversampling}
\alias{generateDesignOversampling}
\title{Generates a random design for a mixed parameter set using oversampling
       and removing "close" design points.}
\usage{
generateDesignOversampling(n = 10L, par.set, trafo = FALSE,
  oversampling = 20L, start.design = generateDesign(1, par.set)[NULL, ])
}
\arguments{
\item{n}{[\code{integer(1)}]\cr
Number of samples in design.
Default is 10.}

\item{par.set}{[\code{\link{ParamSet}}]\cr
Parameter set.}

\item{trafo}{[\code{logical(1)}]\cr
Transform all parameters by using theirs respective transformation functions.
Default is \code{FALSE}.}

\item{oversampling}{[\code{double}]\cr
Oversampling ratio. Must be greater than 1.}

\item{start.design}{[\code{data.frame}]
The start design will be extended until size \code{n} is reached.
The first rows of the generated design will match the start.design.
Default is an empty data.frame.}
}
\value{
[\code{data.frame}]. Columns are named by the ids of the parameters.
  If the \code{par.set} argument contains a vector parameter, its corresponding column names
  in the design are the parameter id concatenated with 1 to dimension of the vector.
  The result will have an \code{logical(1)} attribute \dQuote{trafo},
  which is set to the value of argument \code{trafo}.
}
\description{
The function simply calls \code{\link{generateRandomDesign}} generating
\code{oversampling} more points than desired. Afterwards, repeatedly one of
the 2 points with the smallest Gower distance is removed until the design
is reduced to size \code{n}.
}

