% Generated by roxygen2 (4.0.1): do not edit by hand
\name{getOptPathCol}
\alias{getOptPathCol}
\title{Get column from the optimization path.}
\usage{
getOptPathCol(op, name)
}
\arguments{
\item{op}{[\code{\link{OptPath}}]\cr
Optimization path.}

\item{name}{[\code{character(1)}]\cr
Name of the column.}
}
\value{
Single column as a vector.
}
\description{
Get column from the optimization path.
}
\seealso{
Other optpath: \code{\link{OptPath}},
  \code{\link{OptPathDF}}, \code{\link{makeOptPathDF}};
  \code{\link{addOptPathEl}};
  \code{\link{getOptPathBestIndex}};
  \code{\link{getOptPathCols}};
  \code{\link{getOptPathDOB}}; \code{\link{getOptPathEOL}};
  \code{\link{getOptPathEl}};
  \code{\link{getOptPathErrorMessages}};
  \code{\link{getOptPathExecTimes}};
  \code{\link{getOptPathLength}};
  \code{\link{getOptPathParetoFront}};
  \code{\link{getOptPathY}}; \code{\link{setOptPathElDOB}};
  \code{\link{setOptPathElEOL}}
}

