% Generated by roxygen2 (4.0.0): do not edit by hand
\name{getOptPathDOB}
\alias{getOptPathDOB}
\title{Get date-of-birth vector from the optimization path.}
\usage{
getOptPathDOB(op)
}
\arguments{
\item{op}{[\code{\link{OptPath}}]\cr
Optimization path.}
}
\value{
[\code{integer}].
}
\description{
Get date-of-birth vector from the optimization path.
}

