\name{getOptPathEOL}
\alias{getOptPathEOL}
\title{Get end-of-life vector from the optimization path.}
\usage{
  getOptPathEOL(op)
}
\arguments{
  \item{op}{[\code{\link{OptPath}}]\cr Optimization path.}
}
\value{
  [\code{integer}].
}
\description{
  Get end-of-life vector from the optimization path.
}

