% Generated by roxygen2 (4.0.1): do not edit by hand
\name{getOptPathErrorMessages}
\alias{getOptPathErrorMessages}
\title{Get error-message vector from the optimization path.}
\usage{
getOptPathErrorMessages(op)
}
\arguments{
\item{op}{[\code{\link{OptPath}}]\cr
Optimization path.}
}
\value{
[\code{character}].
}
\description{
Get error-message vector from the optimization path.
}

