% Generated by roxygen2 (4.0.0): do not edit by hand
\name{getOptPathLength}
\alias{getOptPathLength}
\title{Get the length of the optimization path.}
\usage{
getOptPathLength(op)
}
\arguments{
\item{op}{[\code{\link{OptPath}}]\cr
Optimization path.}
}
\value{
[\code{integer(1)}]
}
\description{
Dependent parameters whose requirements are not satisfied are represented by a scalar
NA in the output.
}

