% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getParamSet.R
\name{getParamSet}
\alias{getParamSet}
\title{Extract a ParamSet.}
\usage{
getParamSet(x, ...)
}
\arguments{
\item{x}{[any]\cr
Object to retrieve the \code{\link{ParamSet}} from.}

\item{...}{[any]\cr
Additional arguments. Currently ignored.}
}
\value{
\code{\link{ParamSet}}
}
\description{
This is a generic function used to extract a \code{\link{ParamSet}} out of
 different objects, e.g. an \code{\link{OptPath}}.
}

