% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getParamTypeCounts.R
\name{getParamTypeCounts}
\alias{getParamTypeCounts}
\title{Returns information on the number of parameters of a each type.}
\usage{
getParamTypeCounts(par.set)
}
\arguments{
\item{par.set}{[\code{\link{ParamSet}}]\cr
Parameter set.}
}
\value{
[\code{list}]
 Named list which contains for each supported parameter type the
 number of parameters of this type in the given ParamSet.
}
\description{
Returns information on the number of parameters of a each type.
}
