% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getRequirements.R
\name{getRequirements}
\alias{getRequirements}
\title{Return all require-expressions of a param set.}
\usage{
getRequirements(par.set, no.conditions = "remove")
}
\arguments{
\item{par.set}{[\code{\link{ParamSet}}]\cr
Parameter set.}

\item{no.conditions}{[any] \cr
Params without a requires-setting will result in this value in the returned list.
Default is \dQuote{remove} which means to remove them from the list.}
}
\value{
[named \code{list}].
  Named list of require-call-objects, lengths corresponds to number of params (potentially
  only the subset with requires-field), named with with param ids.
}
\description{
Returns all \code{require}s-objects of a param set as a list.
}

