\name{hasDiscrete}
\alias{hasDiscrete}
\title{Check parameter set for discrete params.}
\usage{
hasDiscrete(par)
}
\arguments{
  \item{par}{\code{\link{ParamSet}}]\cr Parameter set.}
}
\value{
[\code{logical(1)}].
}
\description{
\code{TRUE} iff the parameter set contains at least one
discrete parameter.
}

