% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hasExpression.R
\name{hasExpression}
\alias{hasExpression}
\title{Check if parameter values contain expressions.}
\usage{
hasExpression(par)
}
\arguments{
\item{par}{[\code{\link{Param}} | \code{\link{ParamSet}}]\cr
Parameter or parameter set.}
}
\value{
[\code{logical(1)}].
}
\description{
Check if parameter values contain expressions.
}

