% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hasForbidden.R
\name{hasForbidden}
\alias{hasForbidden}
\title{Check parameter set for forbidden region.}
\usage{
hasForbidden(par.set)
}
\arguments{
\item{par.set}{\link{\code{\link{ParamSet}}}\cr
Parameter set.}
}
\value{
\code{\link{logical(1)}}.
}
\description{
Check parameter set for forbidden region.
}
