\name{hasInteger}
\alias{hasInteger}
\title{Check parameter set for integer params.}
\usage{
hasInteger(par)
}
\arguments{
  \item{par}{\code{\link{ParamSet}}]\cr Parameter set.}
}
\value{
[\code{logical(1)}].
}
\description{
\code{TRUE} iff the parameter set contains at least one
integer parameter.
}

