\name{hasNumeric}
\alias{hasNumeric}
\title{Check parameter set for numeric params.}
\usage{
hasNumeric(par)
}
\arguments{
  \item{par}{\code{\link{ParamSet}}]\cr Parameter set.}
}
\value{
[\code{logical(1)}].
}
\description{
\code{TRUE} iff the parameter set contains at least one
numeric parameter.
}

