% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insertCompliant.R
\name{insertCompliant}
\alias{insertCompliant}
\title{Insert par.vals to old ones with meeting requirements}
\usage{
insertCompliant(old.par.vals, new.par.vals, par.set)
}
\arguments{
\item{old.par.vals}{[\code{list}]\cr
Param Values to be updated.}

\item{new.par.vals}{[\code{list}]\cr
New Param Values to update the \code{old.par.vals}.}

\item{par.set}{[\code{\link{ParamSet}}]\cr
Parameter set.}
}
\value{
[\code{list}]
}
\description{
Update the values of a given parameter setting with a new parameter setting.
  Settings that do not meet the requirements anymore will be deleted from the first given parameter setting.
}

