\name{isPurelyDiscrete}
\alias{isPurelyDiscrete}
\title{Check parameter / parameter set for discrete params.}
\usage{
isPurelyDiscrete(par)
}
\arguments{
  \item{par}{[\code{\link{Param}} |
  \code{\link{ParamSet}}]\cr Parameter or parameter set.}
}
\value{
[\code{logical(1)}].
}
\description{
\code{TRUE} iff the parameter is discrete or all parameters
in the set are discrete.
}

