\name{isPurelyInteger}
\alias{isPurelyInteger}
\title{Check parameter / parameter set for integer params.}
\usage{
isPurelyInteger(par)
}
\arguments{
  \item{par}{[\code{\link{Param}} |
  \code{\link{ParamSet}}]\cr Parameter or parameter set.}
}
\value{
[\code{logical(1)}].
}
\description{
\code{TRUE} iff the parameter is of type integer or all
parameters in the set are integer parameters.
}

