\name{isPurelyNumeric}
\alias{isPurelyNumeric}
\title{Check parameter / parameter set for numeric params.}
\usage{
isPurelyNumeric(par)
}
\arguments{
  \item{par}{[\code{\link{Param}} |
  \code{\link{ParamSet}}]\cr Parameter or parameter set.}
}
\value{
[\code{logical(1)}].
}
\description{
\code{TRUE} iff the parameter is of type numeric or all
parameters in the set are numeric parameters.
}

