% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isSpecialValue.R
\name{isSpecialValue}
\alias{isSpecialValue}
\title{Is a given value in the list of special values for a param?}
\usage{
isSpecialValue(par, x)
}
\arguments{
\item{par}{\link{\code{\link{Param}}}\cr
Parameter.}

\item{x}{\link{any} \cr
Single value to check.}
}
\value{
\code{\link{logical(1)}}.
}
\description{
See title.
}
