% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listToDfOneRow.R
\name{listToDfOneRow}
\alias{listToDfOneRow}
\title{Convert a list to a data.frame with one row}
\usage{
listToDfOneRow(l)
}
\arguments{
\item{l}{(\code{list})\cr
of atomic values of vectors.}
}
\value{
(\code{data.frame}) with only one row, containing the list elements.
}
\description{
Convert a list of vectors or scalars to a \code{data.frame} with only one row. Names of the columns correspond to
the names of elements in the list. If a vector is one list element it is spread over multiple
columns and named sequentially, e.g. \code{a = c(5,7)} becomes \code{data.frame(a1 = 5, a2 = 7)}.
}
