% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsePCSFile.R
\name{parsePCSFile}
\alias{parsePCSFile}
\title{Read and parse PCS files}
\usage{
parsePCSFile(file)
}
\arguments{
\item{file}{[\code{character(1)}].\cr
Path to the PCS file.}
}
\value{
\code{\link{ParamSet}}.
}
