% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/OptPath_plotter.R
\name{plotOptPath}
\alias{plotOptPath}
\title{Plot method for optimization paths.}
\usage{
plotOptPath(op, iters, pause = TRUE, lim.x = list(), lim.y = list(),
  title = "", ...)
}
\arguments{
\item{op}{[\code{OptPath}]\cr
Optimization path.}

\item{iters}{[\code{integer} | NULL]\cr
Vector of iterations which should be plotted one after another. If \code{NULL},
which is the default, all iterations are plotted. Iteration 0 plots
all elements with dob = 0.}

\item{pause}{[\code{logical(1)}]\cr
Should the process be paused after each iteration?
Default is \code{TRUE}.}

\item{lim.x}{[\code{list}],}

\item{lim.y}{[\code{list}]\cr
Axis limits for the plots. Must be a named list, so you can specify the
axis limits for every plot. Every element of the list must be a numeric
vector of length 2. Available names for elements are:
XSpace - limits for the X-Space plot
YSpace - limits for the Y-Space plot
Default is an empty list - in this case limits are automatically set.}

\item{title}{[\code{character(1)} | NULL]\cr
Main title for the arranged plots.}

\item{...}{Additional parameters for \code{\link{renderOptPathPlot}}.}
}
\value{
List of plots, one for each iteration.
}
\description{
Plot method for optimization paths.
}

