% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/OptPath_plotYTraces.R
\name{plotYTraces}
\alias{plotYTraces}
\title{Plots Y traces of multiple optimization paths}
\usage{
plotYTraces(opt.paths)
}
\arguments{
\item{opt.paths}{[\code{list}]\cr
  List of \code{OptPath} objects}
}
\value{
[\code{NULL}]
}
\description{
Plot function for \code{\link{renderYTraces}}
}

