\name{removeMissingValues}
\alias{removeMissingValues}
\title{Removes all scalar NAs from a parameter setting list.}
\usage{
  removeMissingValues(x)
}
\arguments{
  \item{x}{[\code{list}]\cr List of paramter values.}
}
\value{
  [\code{list}].
}
\description{
  Removes all scalar NAs from a parameter setting list.
}

