% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/renderOptPathPlot.R
\name{renderOptPathPlot}
\alias{renderOptPathPlot}
\title{Functions for plotting optimization paths.}
\usage{
renderOptPathPlot(op, iter, xlim = list(), ylim = list(), alpha = TRUE,
  colours = c("red", "blue", "green", "orange"), size.points = 3,
  size.lines = 1.5, impute.scale = 1, impute.value = "missing",
  scale = "std", ggplot.theme = ggplot2::theme(legend.position = "top"),
  marked = NULL, subset.obs, subset.vars, subset.targets, short.x.names,
  short.y.names)
}
\arguments{
\item{op}{[\code{OptPath}]\cr
Optimization path.}

\item{iter}{[\code{integer(1)}]\cr
Selected iteration of \code{x} to render plots for.}

\item{xlim}{[\code{list}]}

\item{ylim}{[\code{list}], \cr
Axis limits for the plots. Must be a named list, so you can specify the
axis limits for every plot. Every element of the list must be a numeric
vector of length 2. Available names for elements are:
XSpace - limits for the X-Space plot
YSpace - limits for the Y-Space plot
Default is an empty list - in this case limits are automatically set.}

\item{alpha}{[\code{logical(1)}]\cr
Activates or deactivates the alpha fading for the plots. Default is \code{TRUE}.}

\item{colours}{[\code{character(4)}]\cr
Colours of the points/lines for the four point types init, seq, prob and marked.
Default is red for init, blue for seq, green for prob and orange for marked.}

\item{size.points}{[\code{numeric(4)} | NULL]\cr
Size of points in the plot, default is 3.}

\item{size.lines}{[\code{numeric(4)} | NULL]\cr
Size of lines in the plots, default is 1.5.}

\item{impute.scale}{[\code{numeric(1)}]\cr
Numeric missing values will be replaced by \code{max + impute.scale * (max - min)}.
Default is \code{1}.}

\item{impute.value}{[\code{character(1)}]\cr
Factor missing values will be replaced by \code{impute.value}. Default is \code{missing}.}

\item{scale}{[\code{character(1)}]\cr
Parameter \code{scale} from the function \code{GGally::ggparcoord} which is used for
 the multiD-case. Default is \code{std}.}

\item{ggplot.theme}{Theme for the ggplots. Can be generated by \code{ggplot2::theme()}. Default is
\code{ggplot2::theme(legend.position = "top")}.}

\item{marked}{[\code{integer} | \code{character(1)} | NULL]\cr
"best" or indices of points that should be marked in the plots.
If \code{marked = "best"} the best point for single crit optimization
respectively the pareto front for multi crit optimization is marked.
Default is \code{NULL} (no points are marked).}

\item{subset.obs}{[\code{integer}]\cr
Vector of indices to subset of observations to be plotted, default is all observations.
All indices must be available in the opt.path. But, to enable subsetting over multiple
iterations, not all indices must be available in the current iteration.
Indices not available in the current iteration will be ignored. Default is all observations.}

\item{subset.vars}{[\code{integer} | \code{character}]\cr
Subset of variables (x-variables) to be plotted. Either vector of indices or names.
Default is all variables}

\item{subset.targets}{[\code{integer} | \code{character}]\cr
Subset of target variables (y-variables) to be plotted. Either vector of indices or names.
Default is all variables}

\item{short.x.names}{[\code{character}],}

\item{short.y.names}{[\code{character}]\cr
Short names for x or y variables that are used as axis labels.}
}
\value{
List of plots. If both X and Y space are 1D, list has length 1,
  otherwise length 2 with one plot for X and Y space respectivly.
}
\description{
Functions for plotting optimization paths.
}

