% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/OptPath_plotYTraces.R
\name{renderYTraces}
\alias{renderYTraces}
\title{Plots Y traces of multiple optimization paths}
\usage{
renderYTraces(opt.paths)
}
\arguments{
\item{opt.paths}{[\code{list}]\cr
List of \code{OptPath} objects}
}
\value{
ggplot2 plot object
}
\description{
Can be used for only single-objective optimization paths.
Useful to compare runs of different algorithms on the same optimization problem.
You can add your own ggplot layers to the resulting plot object.
}

