% Generated by roxygen2 (4.0.1): do not edit by hand
\name{repairPoint}
\alias{repairPoint}
\title{Repairs points outside of box constraints by clipping it to bounds.}
\usage{
repairPoint(par.set, x, warn = FALSE)
}
\arguments{
\item{par.set}{[\code{\link{ParamSet}}]\cr
Parameter set.}

\item{x}{[\code{list}]\cr
List of values which eventually are located outside of the bounds.}

\item{warn}{[\code{logical(1)}]\cr
Boolean indicating whether a warning should be printed each time
a value is being repaired. Default is \code{FALSE}.}
}
\value{
[\code{list}]:
  List of repaired points.
}
\description{
Repairs points outside of box constraints by clipping it to bounds.
}

