% Generated by roxygen2 (4.0.1): do not edit by hand
\name{setOptPathElDOB}
\alias{setOptPathElDOB}
\title{Set the dates of birth of parameter values, in-place.}
\usage{
setOptPathElDOB(op, index, dob)
}
\arguments{
\item{op}{[\code{\link{OptPath}}]\cr
Optimization path.}

\item{index}{[\code{integer}]\cr
Vector of indices of elements.}

\item{dob}{[integer] \cr
Dates of birth, single value or same length of \code{index}.}
}
\value{
Nothing.
}
\description{
Set the dates of birth of parameter values, in-place.
}
\seealso{
Other optpath: \code{\link{OptPath}},
  \code{\link{OptPathDF}}, \code{\link{makeOptPathDF}};
  \code{\link{addOptPathEl}};
  \code{\link{getOptPathBestIndex}};
  \code{\link{getOptPathCol}}; \code{\link{getOptPathDOB}};
  \code{\link{getOptPathEOL}}; \code{\link{getOptPathEl}};
  \code{\link{getOptPathErrorMessages}};
  \code{\link{getOptPathExecTimes}};
  \code{\link{getOptPathLength}};
  \code{\link{getOptPathParetoFront}};
  \code{\link{getOptPathY}}; \code{\link{setOptPathElEOL}}
}

