% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/setValueCNames.R
\name{setValueCNames}
\alias{setValueCNames}
\title{Set components names for vector names}
\usage{
setValueCNames(par, x)
}
\arguments{
\item{par}{[\code{\link{Param}} | \code{\link{ParamSet}}]\cr
Parameter or parameter set.}

\item{x}{[any] \cr
Param value(s).
For a parameter set this must be a list in the correct order.}
}
\value{
\code{x} with changed names.
}
\description{
If param has \code{cnames} set component names in a value.
Otherwise \code{x} is left unchanged.
}

