% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splitVectorParams.R
\name{splitVectorParams}
\alias{splitVectorParams}
\title{Split up a vector param into a list of length-1 params.}
\usage{
splitVectorParams(par)
}
\arguments{
\item{par}{[\code{\link{Param}} | \code{\link{ParamSet}}]\cr
Parameter or parameter set.}
}
\value{
[list of \code{\link{Param}} | \code{\link{ParamSet}}].
  Return a list for single params and a (converted) param set for param sets.
}
\description{
Splits up vector params into a multiple length-1 params.
Sometimes a useful conversion if it is nicer to operate on
the individual normal params.
}
