% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/subsetParams.R
\name{subsetParams}
\alias{subsetParams}
\title{Subset Params from ParamSet by ids.}
\usage{
subsetParams(par.set, subset)
}
\arguments{
\item{par.set}{[\code{\link{ParamSet}}]\cr
Parameter set.}

\item{subset}{[\code{character}]\cr
\code{id}s of the \code{\link{Param}}s in the \code{\link{ParamSet}} to keep from the ParamSet.}
}
\value{
[\code{\link{ParamSet}}].
}
\description{
Subset Params from ParamSet by ids.
}

