% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writePCSFile.R
\name{writePCSFile}
\alias{writePCSFile}
\title{Writes a param set to a PCS files.}
\usage{
writePCSFile(par.set, path, overwrite = FALSE)
}
\arguments{
\item{par.set}{[\code{\link{ParamSet}}]\cr
Parameter set.}

\item{path}{[\code{character(1)}]\cr
File path for output PCS file.}

\item{overwrite}{[\code{logical(1)}]\cr
Can \code{path} be overwritten?
Default is \code{FALSE}.}
}
\value{
Nothing.
}
\description{

}
