% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/calcDistance.R
\name{calcDistance}
\alias{calcDistance}
\title{name calcDistance}
\usage{
calcDistance(par.set, method = "gower", data = NULL, model = NULL)
}
\arguments{
\item{par.set}{[list of \code{\link{ParamSet}} parameters]\cr
Set of parameter configurations, can be a list of configurations
generated from a \code{\link{ParamSet}}, a \code{matrix} or a \code{data.frame}
containing the configurations.}

\item{method}{[\code{character}]\cr
The metric to be used. Currently, one among \code{gower} and \code{randomForest}.
{gower} is the Gower distance, {randomForest} is based on the proximity value computed by a
random forest.\cr}

\item{data\cr}{Data already available.}

\item{model\cr}{A statistical model.\cr}
}
\value{
the distance matrix for \code{par.set}.
}
\description{
name calcDistance
}

