% Generated by roxygen2 (4.0.2): do not edit by hand
\name{dropParams}
\alias{dropParams}
\title{Drop Params form ParamSet by ids.}
\usage{
dropParams(par.set, drop)
}
\arguments{
\item{par.set}{[\code{\link{ParamSet}}]\cr
Parameter set.}

\item{drop}{[\code{character}]\cr
\code{id}s of the \code{\link{Param}}s in the \code{\link{ParamSet}} to drop from the ParamSet.}
}
\value{
[\code{\link{ParamSet}}].
}
\description{
Drop Params form ParamSet by ids.
}

