% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluateParamSet.R
\name{evaluateParamSet}
\alias{evaluateParamSet}
\title{Evaluates expressions of a parameter set.}
\usage{
evaluateParamSet(par.set, dict = NULL)
}
\arguments{
\item{par.set}{[\code{\link{ParamSet}}]\cr
Parameter set.}

\item{dict}{[\code{environment} | \code{list} | \code{NULL}]\cr
Environment or list which will be used for evaluating the variables
of expressions within a parameter set. The default is \code{NULL}.}
}
\value{
[\code{TRUE}].
}
\description{
Evaluates the expressions of a parameter set for a given environment.
}
\examples{
ps = makeParamSet(
  makeNumericParam("x", lower = expression(p), upper = expression(ceiling(3 * p))),
  makeIntegerParam("y", lower = 1, upper = 2)
)
evaluateParamSet(ps, dict = list(p = 3))

ps = makeParamSet(
  makeNumericParam("x", default = expression(sum(data$Species == "setosa"))),
  makeIntegerParam("y", lower = 1, upper = 2),
  keys = c("data", "Species")
)
evaluateParamSet(ps, dict = list(data = iris))
}

