% Generated by roxygen2 (4.0.2): do not edit by hand
\name{extractParamSetFromSooFunction}
\alias{extractParamSetFromSooFunction}
\title{Extracts ParamSet from soobench function.}
\usage{
extractParamSetFromSooFunction(fn)
}
\arguments{
\item{fn}{[\code{\link[soobench]{soo_function}}]\cr
Source function.}
}
\value{
[\code{\link{ParamSet}}]
}
\description{
Extracts ParamSet from soobench function.
}

