% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/getDefaults.R
\name{getDefaults}
\alias{getDefaults}
\title{Return defaults of parameters in parameter set.}
\usage{
getDefaults(par.set, include.null = FALSE)
}
\arguments{
\item{par.set}{[\code{\link{ParamSet}}]\cr
Parameter set.}

\item{include.null}{[\code{logical(1)}]\cr
Include \code{NULL} entries for parameters without default values in the result list?
Note that this can be slightly dangerous as \code{NULL} might be used as default value
for other parameters.
Default is \code{FALSE}.}
}
\value{
[named \code{list}]. Named and in same order as \code{par.set}.
  Parameters without defaults are not present in the list.
}
\description{
Return defaults of parameters in parameter set.
}

