% Generated by roxygen2 (4.0.2): do not edit by hand
\name{getDefaults}
\alias{getDefaults}
\title{Return defaults of parameters in parameter set.}
\usage{
getDefaults(par.set)
}
\arguments{
\item{par.set}{[\code{\link{ParamSet}}]\cr
Parameter set.}
}
\value{
[named \code{list}]. Named and in same order as \code{par.set}.
  Parameters without defaults are not present in the list.
}
\description{
Return defaults of parameters in parameter set.
}

