% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/OptPath_getter.R
\name{getOptPathCol}
\alias{getOptPathCol}
\title{Get column from the optimization path.}
\usage{
getOptPathCol(op, name, dob, eol)
}
\arguments{
\item{op}{[\code{\link{OptPath}}]\cr
Optimization path.}

\item{name}{[\code{character(1)}]\cr
Name of the column.}

\item{dob}{[\code{integer}]\cr
Vector of date-of-birth values to further subset the result.
Only elements with a date-of-birth included in \code{dob} are selected.
Default is all.}

\item{eol}{[\code{integer}]\cr
Vector of end-of-life values to further subset the result.
Only elements with an end-of-life included in \code{eol} are selected.
Default is all.}
}
\value{
Single column as a vector.
}
\description{
Get column from the optimization path.
}
\seealso{
Other optpath: \code{\link{OptPath}},
  \code{\link{OptPathDF}}, \code{\link{makeOptPathDF}};
  \code{\link{addOptPathEl}};
  \code{\link{getOptPathBestIndex}};
  \code{\link{getOptPathCols}};
  \code{\link{getOptPathDOB}}; \code{\link{getOptPathEOL}};
  \code{\link{getOptPathEl}};
  \code{\link{getOptPathErrorMessages}};
  \code{\link{getOptPathExecTimes}};
  \code{\link{getOptPathLength}};
  \code{\link{getOptPathParetoFront}};
  \code{\link{getOptPathX}}; \code{\link{getOptPathY}};
  \code{\link{setOptPathElDOB}};
  \code{\link{setOptPathElEOL}}
}

