% Generated by roxygen2 (4.0.1): do not edit by hand
\name{getOptPathCols}
\alias{getOptPathCols}
\title{Get columns the optimization path. If}
\usage{
getOptPathCols(op, names, check.names = TRUE)
}
\arguments{
\item{op}{[\code{\link{OptPath}}]\cr
Optimization path.}

\item{names}{[\code{character}]\cr
Vector of names of columns}

\item{check.names}{[\code{logical(1)}]
Should the existence of the names be checked? If FALSE NULL is returned for
non-existent cols.}
}
\value{
[\code{data.frame}].
}
\description{
Get columns the optimization path. If
}
\seealso{
Other optpath: \code{\link{OptPath}},
  \code{\link{OptPathDF}}, \code{\link{makeOptPathDF}};
  \code{\link{addOptPathEl}};
  \code{\link{getOptPathBestIndex}};
  \code{\link{getOptPathDOB}}; \code{\link{getOptPathEOL}};
  \code{\link{getOptPathEl}};
  \code{\link{getOptPathErrorMessages}};
  \code{\link{getOptPathExecTimes}};
  \code{\link{getOptPathLength}};
  \code{\link{getOptPathParetoFront}};
  \code{\link{getOptPathY}}; \code{\link{setOptPathElDOB}};
  \code{\link{setOptPathElEOL}}
}

