% Generated by roxygen2 (4.0.1): do not edit by hand
\name{getOptPathCols}
\alias{getOptPathCols}
\title{Get columns from the optimization path.}
\usage{
getOptPathCols(op, names, row.names = NULL)
}
\arguments{
\item{op}{[\code{\link{OptPath}}]\cr
Optimization path.}

\item{names}{[\code{character}]\cr
Names of the columns.}

\item{row.names}{[\code{character}]\cr
Row names for result.
Default is none.}
}
\value{
[\code{data.frame}].
}
\description{
Get columns from the optimization path.
}
\seealso{
Other optpath: \code{\link{OptPath}},
  \code{\link{OptPathDF}}, \code{\link{makeOptPathDF}};
  \code{\link{addOptPathEl}};
  \code{\link{getOptPathBestIndex}};
  \code{\link{getOptPathCol}}; \code{\link{getOptPathDOB}};
  \code{\link{getOptPathEOL}}; \code{\link{getOptPathEl}};
  \code{\link{getOptPathErrorMessages}};
  \code{\link{getOptPathExecTimes}};
  \code{\link{getOptPathLength}};
  \code{\link{getOptPathParetoFront}};
  \code{\link{getOptPathY}}; \code{\link{setOptPathElDOB}};
  \code{\link{setOptPathElEOL}}
}

