% Generated by roxygen2 (4.0.1): do not edit by hand
\name{getOptPathEOL}
\alias{getOptPathEOL}
\title{Get end-of-life vector from the optimization path.}
\usage{
getOptPathEOL(op)
}
\arguments{
\item{op}{[\code{\link{OptPath}}]\cr
Optimization path.}
}
\value{
[\code{integer}].
}
\description{
Get end-of-life vector from the optimization path.
}
\seealso{
Other optpath: \code{\link{OptPath}},
  \code{\link{OptPathDF}}, \code{\link{makeOptPathDF}};
  \code{\link{addOptPathEl}};
  \code{\link{getOptPathBestIndex}};
  \code{\link{getOptPathCols}};
  \code{\link{getOptPathCol}}; \code{\link{getOptPathDOB}};
  \code{\link{getOptPathEl}};
  \code{\link{getOptPathErrorMessages}};
  \code{\link{getOptPathExecTimes}};
  \code{\link{getOptPathLength}};
  \code{\link{getOptPathParetoFront}};
  \code{\link{getOptPathY}}; \code{\link{setOptPathElDOB}};
  \code{\link{setOptPathElEOL}}
}

