% Generated by roxygen2 (4.0.1): do not edit by hand
\name{getOptPathErrorMessages}
\alias{getOptPathErrorMessages}
\title{Get error-message vector from the optimization path.}
\usage{
getOptPathErrorMessages(op)
}
\arguments{
\item{op}{[\code{\link{OptPath}}]\cr
Optimization path.}
}
\value{
[\code{character}].
}
\description{
Get error-message vector from the optimization path.
}
\seealso{
Other optpath: \code{\link{OptPath}},
  \code{\link{OptPathDF}}, \code{\link{makeOptPathDF}};
  \code{\link{addOptPathEl}};
  \code{\link{getOptPathBestIndex}};
  \code{\link{getOptPathCol}}; \code{\link{getOptPathDOB}};
  \code{\link{getOptPathEOL}}; \code{\link{getOptPathEl}};
  \code{\link{getOptPathExecTimes}};
  \code{\link{getOptPathLength}};
  \code{\link{getOptPathParetoFront}};
  \code{\link{getOptPathY}}; \code{\link{setOptPathElDOB}};
  \code{\link{setOptPathElEOL}}
}

