% Generated by roxygen2 (4.0.0): do not edit by hand
\name{getOptPathY}
\alias{getOptPathY}
\title{Get y-vector or y-matrix from the optimization path.}
\usage{
getOptPathY(op, names, drop = TRUE)
}
\arguments{
\item{op}{[\code{\link{OptPath}}]\cr
Optimization path.}

\item{names}{[\code{character}]\cr
Names of performance measure.
Default is all performance measures in path.}

\item{drop}{[\code{logical(1)}]\cr
Return vector instead of matrix when only one y-column was selected?
Default is \code{TRUE}.}
}
\value{
[\code{numeric} | \code{matrix}]. The columns of the matrix are always named.
}
\description{
Get y-vector or y-matrix from the optimization path.
}

