\name{getOptPathY}
\alias{getOptPathY}
\title{Get y vector from the optimization path.}
\usage{
  getOptPathY(op, name)
}
\arguments{
  \item{op}{[\code{\link{OptPath}}]\cr Optimization path.}

  \item{name}{[\code{character(1)}]\cr Name of performance
  measure.}
}
\value{
  [\code{numeric}].
}
\description{
  Get y vector from the optimization path.
}

