% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getParamSet.R
\name{getParamSet}
\alias{getParamSet}
\title{Get parameter set.}
\usage{
getParamSet(x)
}
\arguments{
\item{x}{[\code{object}] \cr
Object to extract the ParamSet from.}
}
\value{
[\code{\link[ParamHelpers]{ParamSet}}]
}
\description{
\code{getParamSet} is a generic and can be called to extract the \code{ParamSet} from different objects.
}

