% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getRequirements.R
\name{getRequirements}
\alias{getRequirements}
\title{Return all require-expressions of a param set.}
\usage{
getRequirements(par.set, remove.null = TRUE)
}
\arguments{
\item{par.set}{\link{ParamSet}\cr
Parameter set.}

\item{remove.null}{(\code{logical(1)})\cr
If not set, params without a requires-setting will result in a \code{NULL}
element in the returned list, otherwise they are removed. Default is
code{TRUE}.}
}
\value{
xnamed \code{list}.
Named list of require-call-objects, lengths corresponds to number of params
(potentially only the subset with requires-field), named with with param
ids.
}
\description{
Returns all \code{require}s-objects of a param set as a list.
}
