% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTypeStrings.R
\name{getTypeStrings}
\alias{getTypeStrings}
\alias{getTypeStringsDiscrete}
\alias{getTypeStringsInteger}
\alias{getTypeStringsNumeric}
\alias{getTypeStringsNumericStrict}
\title{Get parameter type-strings.}
\usage{
getTypeStringsNumeric(include.int = TRUE)

getTypeStringsNumericStrict()

getTypeStringsInteger()

getTypeStringsDiscrete(include.logical = TRUE)
}
\arguments{
\item{include.int}{[\code{logical(1)}]\cr
Are integers also considered to be numeric?
Default is \code{TRUE}.}

\item{include.logical}{[\code{logical(1)}]\cr
Are logicals also considered to be discrete?
Default is \code{TRUE}.}
}
\value{
[\code{character}].
}
\description{
Returns type strings used in \code{param$type} for certain groups of parameters.
}

