% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hasFiniteBoxConstraints.R
\name{hasFiniteBoxConstraints}
\alias{hasFiniteBoxConstraints}
\title{Checks if a parameter or each parameter of a parameter set has ONLY finite lower and upper bounds.}
\usage{
hasFiniteBoxConstraints(par, dict = NULL)
}
\arguments{
\item{par}{[\code{\link{Param}} | \code{\link{ParamSet}}]\cr
Parameter or parameter set.}

\item{dict}{[\code{environment} | \code{list} | \code{NULL}]\cr
Environment or list which will be used for evaluating the variables
of expressions within a parameter, parameter set or list of parameters.
The default is \code{NULL}.}
}
\value{
[\code{logical(1)}]
}
\description{
Checks if a parameter or each parameter of a parameter set has ONLY finite lower and upper bounds.
}
