% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hasRequires.R
\name{hasRequires}
\alias{hasRequires}
\title{Check parameter / parameter set for requirements / dependencies.}
\usage{
hasRequires(par)
}
\arguments{
\item{par}{(\link{Param} | \link{ParamSet})\cr
Parameter or parameter set.}
}
\value{
\code{logical(1)}.
}
\description{
\code{TRUE} iff the parameter has any requirements or any parameter in the set has
requirements.
}
