% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hasTrafo.R
\name{hasTrafo}
\alias{hasTrafo}
\title{Check parameter / parameter set for trafos.}
\usage{
hasTrafo(par)
}
\arguments{
\item{par}{[\code{\link{Param}} | \code{\link{ParamSet}}]\cr
Parameter or parameter set.}
}
\value{
[\code{logical(1)}].
}
\description{
\code{TRUE} iff the parameter has any trafos or any parameter in the set has
trafos.
}

