% Generated by roxygen2 (4.0.1): do not edit by hand
\name{hasTrafo}
\alias{hasTrafo}
\title{Check parameter / parameter set for trafos.}
\usage{
hasTrafo(par)
}
\arguments{
\item{par}{[\code{\link{Param}} | \code{\link{ParamSet}}]\cr
Parameter or parameter set.}
}
\value{
[\code{logical(1)}].
}
\description{
\code{TRUE} iff the parameter has any trafos or any parameter in the set has
trafos.
}

