% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ParamSet.R
\name{isEmpty}
\alias{isEmpty}
\title{Check whether parameter set is empty.}
\usage{
isEmpty(par.set)
}
\arguments{
\item{par.set}{\code{\link{ParamSet}}]\cr
Parameter set.}
}
\value{
[\code{logical(1)}].
}
\description{
Check whether parameter set is empty.
}
