% Generated by roxygen2 (4.0.2): do not edit by hand
\name{isForbidden}
\alias{isForbidden}
\title{Check whether parameter setting lies in forbidden region of parameter set.}
\usage{
isForbidden(par.set, x)
}
\arguments{
\item{par.set}{[\code{\link{ParamSet}}]\cr
Parameter set.}

\item{x}{[named \code{list}] \cr
Parameter setting to check.}
}
\value{
[\code{logical(1)}].
}
\description{
Parameter sets without a forbidden region always return \code{FALSE}.
}

