% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isSpecialValue.R
\name{isSpecialValue}
\alias{isSpecialValue}
\title{Is a given value in the list of special values for a param?}
\usage{
isSpecialValue(par, x)
}
\arguments{
\item{par}{[\code{\link{Param}}]\cr
Parameter.}

\item{x}{[any] \cr
Single value to check.}
}
\value{
[\code{logical(1)}].
}
\description{
See title.
}
