% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isType.R
\name{isType}
\alias{isType}
\alias{isNumeric}
\alias{isDiscrete}
\alias{isInteger}
\alias{isLogical}
\alias{isCharacter}
\title{Check parameter / parameter set contain ONLY a certain type.}
\usage{
isNumeric(par, include.int = TRUE)

isDiscrete(par, include.logical = TRUE)

isInteger(par)

isLogical(par)

isCharacter(par)
}
\arguments{
\item{par}{[\code{\link{Param}} | \code{\link{ParamSet}}]\cr
Parameter or parameter set.}

\item{include.int}{[\code{logical(1)}]\cr
Are integers also considered to be numeric?
Default is \code{TRUE}.}

\item{include.logical}{[\code{logical(1)}]\cr
Are logicals also considered to be discrete?
Default is \code{TRUE}.}
}
\value{
[\code{logical(1)}]
}
\description{
An empty param set is considered to be of all types.
}
