% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isVector.R
\name{isVector}
\alias{isVector}
\title{Check parameter / parameter set for vector params.}
\usage{
isVector(par)
}
\arguments{
\item{par}{(\link{Param} | \link{ParamSet})\cr
Parameter or parameter set.}
}
\value{
\code{logical(1)}.
}
\description{
\code{TRUE} if the parameter is a vector parameter or all parameters in the
set are vector parameters.
}
