% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/OptPath_plotter.R
\name{plotOptPath}
\alias{plotOptPath}
\title{Plot method for optimization paths.}
\usage{
plotOptPath(op, iters, pause = TRUE, alpha = TRUE, lim.x = NULL,
  lim.y = NULL)
}
\arguments{
\item{op}{[\code{OptPath}]\cr
Optimization path.}

\item{iters}{[\code{integer} | NULL]\cr
Vector of iteration which should be plotted one after another. If \code{NULL},
which is the default, all iterations are plotted. Iteration 0 plots
all elements with dob = 0}

\item{pause}{[\code{logical(1)}]\cr
Should the process be paused after each iteration?
Default is \code{TRUE}.}

\item{alpha}{[\code{logical(1)}]\cr
Activates or deactivates the alpha fading for the parallel X-space plot. Default is \code{TRUE}.}

\item{lim.x}{[\code{list} | NULL]\cr
Length of list equals dimensionality of X-Space. Not used, if the dimension
is higher than 2. Each element is a \code{numeric(2)}, determining
the x limits for the plots. Default is NULL, in this case min and max
values of the data will be used.}

\item{lim.y}{[\code{list} | NULL]\cr
Length of list equals dimensionality of X-Space. Not used, if the dimension
is higher than 2. Each element is a \code{numeric(2)}, determining
the y limits for the plots. Default is NULL, in this case min and max
values of the data will be used.}
}
\value{
List of plots, one for each iteration.
}
\description{
Plot method for optimization paths.
}

