% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/OptPath_plotter.R
\name{plotOptPath}
\alias{plotOptPath}
\title{Plot method for optimization paths.}
\usage{
plotOptPath(op, iters, pause = TRUE, alpha = TRUE, lim.x = list(),
  lim.y = list(), title = "", colours = c("red", "blue", "green",
  "orange"), size = c(3, 1.5), impute.scale = 1, impute.value = "missing",
  scale = "std", ggplot.theme = ggplot2::theme(legend.position = "top"),
  marked = NULL)
}
\arguments{
\item{op}{[\code{OptPath}]\cr
Optimization path.}

\item{iters}{[\code{integer} | NULL]\cr
Vector of iteration which should be plotted one after another. If \code{NULL},
which is the default, all iterations are plotted. Iteration 0 plots
all elements with dob = 0.}

\item{pause}{[\code{logical(1)}]\cr
Should the process be paused after each iteration?
Default is \code{TRUE}.}

\item{alpha}{[\code{logical(1)}]\cr
Activates or deactivates the alpha fading for the plots. Default is \code{TRUE}.}

\item{lim.x}{[\code{list}],}

\item{lim.y}{[\code{list}]\cr
Axis limits for the plots. Must be a named list, so you can specify the
axis limits for every plot. Every element of the list must be a numeric
vector of length 2. Available names for elements are:
XSpace - limits for the X-Space plot
YSpace - limits for the Y-Space plot
Default is an empty list - in this case limits are automatically set.}

\item{title}{[\code{character} | NULL]\cr
Main title for the arranged plots.}

\item{colours}{[\code{character(4)}]\cr
Colours of the points/lines for the three point types init, seq, prob and marked.}

\item{size}{[\code{numeric(2)} | NULL]\cr
Size of points (1st entry of \code{size}) or lines (2nd entry of \code{size}).
The default is \code{c(3, 1.5)}.}

\item{impute.scale}{[\code{numeric(1)}]\cr
Numeric missing values will be replaced by \code{max + impute.scale * (max - min)}.
Default is \code{1}.}

\item{impute.value}{[\code{character(1)}]\cr
Factor missing values will be replaced by \code{impute.value}. Default is \code{missing}.}

\item{scale}{[\code{character(1)}]\cr
Parameter \code{scale} from the function \code{GGally::ggparcoord} which is used for
 the multiD-case. Default is \code{std}.}

\item{ggplot.theme}{Theme for the ggplots. Can be generated by \code{ggplot2::theme()}. Default is
\code{ggplot2::theme(legend.position = "top")}.}

\item{marked}{[\code{integer()}]\cr
Indices of points that should be marked in the plots. Default is \code{NULL}
(no points are marked).}
}
\value{
List of plots, one for each iteration.
}
\description{
Plot method for optimization paths.
}

