% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OptPath_plotYTraces.R
\name{plotYTraces}
\alias{plotYTraces}
\title{Plots Y traces of multiple optimization paths}
\usage{
plotYTraces(opt.paths, over.time = "dob")
}
\arguments{
\item{opt.paths}{\code{\link{list}}\cr
List of \code{OptPath} objects}

\item{over.time}{\code{\link{character}}\cr
Should the traces be plotted versus the iteration number or the cumulated
execution time? For the later, the opt.path has to contain a extra column
names exec.time. Possible values are dob and exec.time, default is \code{dob}.}
}
\value{
\code{\link{NULL}}
}
\description{
Plot function for \code{\link[=renderYTraces]{renderYTraces()}}
}
