% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/renderOptPathPlot.R
\name{renderOptPathPlot}
\alias{renderOptPathPlot}
\title{Functions for plotting optimization paths.}
\usage{
renderOptPathPlot(op, iter, alpha = TRUE, lim.x = list(), lim.y = list())
}
\arguments{
\item{op}{[\code{OptPath}]\cr
Optimization path.}

\item{iter}{[\code{integer(1)} | NULL]\cr
Selected iteration of \code{x} to render plots for.}

\item{alpha}{[\code{logical(1)}]\cr
Activates or deactivates the alpha fading for the parallel X-space plot. Default is \code{TRUE}.}

\item{lim.x}{[\code{list}],}

\item{lim.y}{[\code{list}]\cr
Axis limits for the plots. Must be a named list, so you can specify the
axis limits for every plot. Every element of the list must be a numeric
vector of length 2. Available names for elements are:
XSpace - limits for the X-Space plot
YSpace - limits for the Y-Space plot
Default is an empty list - in this case limits are automatically set.}
}
\value{
List of plots. If both X and Y space are 1D, list has length 1,
  otherwise length 2 with one plot for X and Y space respectivly.
}
\description{
Functions for plotting optimization paths.
}

