% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/renderOptPathPlot.R
\name{renderOptPathPlot}
\alias{renderOptPathPlot}
\title{Plot method for optimization paths.}
\usage{
renderOptPathPlot(op, iters = NULL, alpha = TRUE)
}
\arguments{
\item{op}{[\code{OptPath}]\cr
Optimization path.}

\item{iters}{[\code{numeric} | NULL]\cr
Vector of iteration which should be plotted one after another. If \code{NULL},
which is the default, all iterations are plotted.}

\item{alpha}{[\code{logical(1)}]\cr
Activates or deactivates the alpha fading for the parallel X-space plot. Default is \code{TRUE}.}
}
\value{
List of plots, one for each iteration.
}
\description{
Plot method for optimization paths.
}

