% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/renderOptPathPlot.R
\name{renderOptPathPlot}
\alias{renderOptPathPlot}
\title{Functions for plotting optimization paths.}
\usage{
renderOptPathPlot(op, iter, lim.x = list(), lim.y = list(), alpha = TRUE,
  colours = c("red", "blue", "green", "orange"), size = c(3, 1.5),
  impute.scale = 1, impute.value = "missing", scale = "std",
  ggplot.theme = ggplot2::theme(legend.position = "top"), marked = NULL,
  subset.obs, subset.vars, subset.targets)
}
\arguments{
\item{op}{[\code{OptPath}]\cr
Optimization path.}

\item{iter}{[\code{integer(1)} | NULL]\cr
Selected iteration of \code{x} to render plots for.}

\item{lim.x}{[\code{list}],}

\item{lim.y}{[\code{list}]\cr
Axis limits for the plots. Must be a named list, so you can specify the
axis limits for every plot. Every element of the list must be a numeric
vector of length 2. Available names for elements are:
XSpace - limits for the X-Space plot
YSpace - limits for the Y-Space plot
Default is an empty list - in this case limits are automatically set.}

\item{alpha}{[\code{logical(1)}]\cr
Activates or deactivates the alpha fading for the plots. Default is \code{TRUE}.}

\item{colours}{[\code{character(4)}]\cr
Colours of the points/lines for the three point types init, seq, prob and marked.}

\item{size}{[\code{numeric(2)} | NULL]\cr
Size of points (1st entry of \code{size}) or lines (2nd entry of \code{size}).
The default is \code{c(3, 1.5)}.}

\item{impute.scale}{[\code{numeric(1)}]\cr
Numeric missing values will be replaced by \code{max + impute.scale * (max - min)}.
Default is \code{1}.}

\item{impute.value}{[\code{character(1)}]\cr
Factor missing values will be replaced by \code{impute.value}. Default is \code{missing}.}

\item{scale}{[\code{character(1)}]\cr
Parameter \code{scale} from the function \code{GGally::ggparcoord} which is used for
 the multiD-case. Default is \code{std}.}

\item{ggplot.theme}{Theme for the ggplots. Can be generated by \code{ggplot2::theme()}. Default is
\code{ggplot2::theme(legend.position = "top")}.}

\item{marked}{[\code{integer} | \code{character(1)} | NULL]\cr
"min" or "max" or indices of points that should be marked in the plots.
Default is \code{NULL} (no points are marked).}

\item{subset.obs}{[\code{integer}]\cr
Vector of indices to subset of observations to be plotted, default is all observations.
All indices must be available in the opt.path, but not in the current iteration.
Indices not available in the current iteration will be ignored.}

\item{subset.vars}{[\code{integer} | \code{character}]\cr
Subset of variables (x-variables) to be plotted. Either vector of indices or names.
Default is all variables}

\item{subset.targets}{[\code{integer} | \code{character}]\cr
Subset of target variables (y-variables) to be plotted. Either vector of indices or names.
Default is all variables}
}
\value{
List of plots. If both X and Y space are 1D, list has length 1,
  otherwise length 2 with one plot for X and Y space respectivly.
}
\description{
Functions for plotting optimization paths.
}

