% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OptPath_plotYTraces.R
\name{renderYTraces}
\alias{renderYTraces}
\title{Plots Y traces of multiple optimization paths}
\usage{
renderYTraces(opt.paths, over.time = "dob")
}
\arguments{
\item{opt.paths}{[\code{\link[=OptPath]{OptPath()}} | list of \code{\link[=OptPath]{OptPath()}}]\cr
Object(s) to plot.}

\item{over.time}{\code{\link{character}}\cr
Should the traces be plotted versus the iteration number or the cumulated
execution time? For the later, the opt.path has to contain a extra column
names exec.time. Possible values are dob and exec.time, default is \code{dob}.}
}
\value{
ggplot2 plot object
}
\description{
Can be used for only single-objective optimization paths. Useful
to compare runs of different algorithms on the same optimization problem. You
can add your own ggplot layers to the resulting plot object.
}
