% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repairPoint.R
\name{repairPoint}
\alias{repairPoint}
\title{Repairs values of numeric and integer parameters out side of constraints.}
\usage{
repairPoint(par.set, x, warn = FALSE)
}
\arguments{
\item{par.set}{[\code{\link{ParamSet}}]\cr
Parameter set.}

\item{x}{[\code{list}]\cr
List of parameter values. Must be in correct order.
Values corresponding to non-numeric/integer types are left unchanged.}

\item{warn}{[\code{logical(1)}]\cr
Boolean indicating whether a warning should be printed each time a value is repaired.
Default is \code{FALSE}.}
}
\value{
[\code{list}]:
  List of repaired points.
}
\description{
Clips values outside of box constraints to bounds.
}

