% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OptPath_setter.R
\name{setOptPathElDOB}
\alias{setOptPathElDOB}
\title{Set the dates of birth of parameter values, in-place.}
\usage{
setOptPathElDOB(op, index, dob)
}
\arguments{
\item{op}{\link{OptPath}\cr
Optimization path.}

\item{index}{\link{integer}\cr
Vector of indices of elements.}

\item{dob}{\link{integer} \cr
Dates of birth, single value or same length of \code{index}.}
}
\value{
Nothing.
}
\description{
Set the dates of birth of parameter values, in-place.
}
\seealso{
Other optpath: 
\code{\link{OptPath}},
\code{\link{addOptPathEl}()},
\code{\link{getOptPathBestIndex}()},
\code{\link{getOptPathCols}()},
\code{\link{getOptPathCol}()},
\code{\link{getOptPathDOB}()},
\code{\link{getOptPathEOL}()},
\code{\link{getOptPathEl}()},
\code{\link{getOptPathErrorMessages}()},
\code{\link{getOptPathExecTimes}()},
\code{\link{getOptPathLength}()},
\code{\link{getOptPathParetoFront}()},
\code{\link{getOptPathX}()},
\code{\link{getOptPathY}()},
\code{\link{setOptPathElEOL}()}
}
\concept{optpath}
