\name{setOptPathElDOB}
\alias{setOptPathElDOB}
\title{Set the dates of birth of parameter values, in-place.}
\usage{
  setOptPathElDOB(op, index, dob)
}
\arguments{
  \item{op}{[\code{\link{OptPath}}]\cr Optimization path.}

  \item{index}{[\code{integer}]\cr Vector of indices of
  elements.}

  \item{dob}{[integer] \cr Dates of birth, single value or
  same length of \code{index}.}
}
\value{
  Nothing.
}
\description{
  Set the dates of birth of parameter values, in-place.
}

